feat(mev-arbitrum): multi-leg encoders + sandwich strategy + V3/V4 dex-kind fix
Adapter capability:
- AaveV3FlashLoanBuilder + MorphoFlashLoanBuilder now encode matchInternal
  (selector 5f188678) and composeFourLeg (selector 72c0469b) calldata in
  addition to executeNativeArb. Selectors match the canonical signatures in
  rust/src/execution.rs.
- Rust execution_py.rs exposes encode_match_internal_calldata and
  encode_compose_four_leg_calldata via the degenbot_rs PyO3 binding.

Strategies:
- New strategies_coordinator/sandwich.py: traditional offensive sandwich
  (V2 reserves, integer solve_v2_sandwich, NativeArbParams round-trip,
  env-flag gated via strategy_sandwich_enabled).
- decision/precedence.py + types.py: 'sandwich' DecisionRoute kind added.
- strategies_coordinator/oracle_sandwich.py: _kind_to_dex_kind now maps
  V3 -> DEX_KIND.V3 and V4 -> DEX_KIND.V4 (previously defaulted to V2 for
  all kinds).

Type-safety + hygiene:
- simulation.py Simulator.__init__ uses direct typed import of RevmDb
  (restores stub visibility against degenbot_rs.pyi).
- provider/__init__.py removes redundant cast() on AlloyProvider;
  provider/interface.py adds cast() on web3 returns to narrow loose types.
- connection/ipc.py renames msg -> payload to fix variable shadowing where
  input was overwritten by local error strings.
- strategy_signals/ostium_oracle_gap.py renames payload fields to snake_case
  with Field(alias=...) + populate_by_name=True (wire compat preserved).
- stylus/pool_adapter/src/lib.rs gates read_int24_from_3bytes with
  cfg(not(any(test, feature='native-test'))) consistent with the rest of
  impl PoolAdapter.

Tests:
- test_aave_v3_flashloan_adapter.py + test_morpho_flashloan_adapter.py
  assert function selectors on the matchInternal + composeFourLeg success
  paths.
- New test_production_blockers.py covers the sandwich DecisionRoute kind
  and flash-route wrap defaults.
- New repro_issue_0057.py covers the Aave debug logger JSONL output.

Gitignore:
- receipt.json runtime artifact suppressed.

Conformance: docs/runbooks/conformance/2026-05-31-degenbot-submodule-diff.md
(CLEAN against 00A/00B doctrine and ADRs 007/009/016/021/026/027/029/030).

+++ src/degenbot/matching/price_compat.py
@@ -23,8 +23,8 @@ def normalize_amount_to_wad(amount: int, decimals: int) -> int:
     if decimals == 18:
         return amount
     if decimals < 18:
-        return amount * 10 ** (18 - decimals)
-    return amount // 10 ** (decimals - 18)
+        return int(amount * 10 ** (18 - decimals))
+    return int(amount // 10 ** (decimals - 18))
 
 
 def denormalize_amount_from_wad(amount_wad: int, decimals: int) -> int:
@@ -36,8 +36,8 @@ def denormalize_amount_from_wad(amount_wad: int, decimals: int) -> int:
     if decimals == 18:
         return amount_wad
     if decimals < 18:
-        return amount_wad // 10 ** (18 - decimals)
-    return amount_wad * 10 ** (decimals - 18)
+        return int(amount_wad // 10 ** (18 - decimals))
+    return int(amount_wad * 10 ** (decimals - 18))
 
 
 def project_sell_to_buy(
